# backend/internal/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HRIS LDP API
  version: 1.0.0
  description: Dokumentasi endpoint utama HRIS LDP.
servers:
  - url: /api
    description: API endpoints
  - url: /
    description: Public/root endpoints
...
paths:
...
  /admin/dashboard:
    get:
      tags: [Dashboard]
      summary: Ambil data /admin/dashboard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
  /admin-staff/dashboard:
    get:
      tags: [AdminStaff]
      summary: Ambil data /admin-staff/dashboard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
  /dashboard:
    get:
      tags: [Dashboard]
      summary: Ambil data /dashboard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
  /pelamar/dashboard:
    get:
      tags: [Pelamar]
      summary: Ambil data /pelamar/dashboard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
  /pelamar/profil:
    get:
      tags: [Pelamar]
      summary: Ambil data /pelamar/profil
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
    post:
      tags: [Pelamar]
      summary: Proses request /pelamar/profil
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
  /staff/dashboard:
    get:
      tags: [Staff]
      summary: Ambil data /staff/dashboard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
  /super-admin/dashboard:
    get:
      tags: [SuperAdmin]
      summary: Ambil data /super-admin/dashboard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
  /super-admin/kelola-divisi:
    get:
      tags: [SuperAdmin]
      summary: Ambil data /super-admin/kelola-divisi
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
    post:
      tags: [SuperAdmin]
      summary: Proses request /super-admin/kelola-divisi
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthenticated
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: hris_session
...